fix: merge grpc_i18n.go into yak.go for CI single-file build compatibility
CI workflows build with 'go build common/yak/cmd/yak.go' (single file),
not './common/yak/cmd/' (directory). The separate grpc_i18n.go was not
included in the build, causing 'undefined: grpcEventPhaseI18n' errors.
Merge all i18n maps and helper functions into yak.go, delete grpc_i18n.go.
